+++ R/collapse_range.R
@@ -19,8 +19,9 @@
 #' \code{original community data}
 #'
 #' @references
-#' Daru, B.H., Farooq, H., Antonelli, A. & Faurby, S. Endemism patterns are
-#' scale dependent.
+#' Daru, B.H., Farooq, H., Antonelli, A. & Faurby, S. (2020) Endemism
+#' patterns are scale dependent. \emph{Nature Communications}
+#' \strong{11}: 2115.
 #' @examples
 #' library(ape)
 #' tr1 <- read.tree(text ="(((a:2,(b:1,c:1):1):1,d:3):1,e:4);")
@@ -44,7 +45,6 @@ collapse_range <- function(x, tree, n, species="species", grids="grids",
     x <- as.data.frame(x)}
   else if (format == "long") {
     x <- as.data.frame(x)
-    x <- x[, c(grids, species)]
     x <- x[, c(grids, species)]
     names(x) <- c("grids", "species")}
   ind <- intersect(x$species, tree$tip.label)