+++ NEWS.md
@@ -1,4 +1,13 @@
-# Changes in version 1.0.3
+# [April 30, 2021] Changes in `phyloregion` version 1.0.5
+
+* Revised `plot_swatch` to use `col` instead of palette argument.
+
+* Added new function, `sdm`, for species distribution modelling.
+
+* Improved functions for transforming converting input data.
+
+
+# [July 19, 2020] Changes in `phyloregion` version 1.0.3
 
 * Removed `plot_evoldistinct` because the functionality is replaced 
 with `plot_swatch`.
@@ -12,6 +21,6 @@ regionalization and macroecology"
 * Improvements in the vignettes with several examples.
 
 
-# phyloregion 1.0.0
+# [March 30, 2020] `phyloregion` version 1.0.0
 
 * This is the first CRAN release of phyloregion.